+++ app/(auth)/login/page.tsx
@@ -1,10 +1,10 @@
 import { Metadata } from "next"
 
-import LoginPage from "@/modules/auth/presentation/pages/default/login-page"
+import DefaultLoginPage from "@/modules/login/presentation/pages/default/DefaultLoginPage"
 
 export const metadata: Metadata = {
   title: "Login",
   description: "Login to your account",
 }
 
-export default LoginPage
+export default DefaultLoginPage
