+++ app/(auth)/register/page.tsx
@@ -1,10 +1,10 @@
 import { Metadata } from "next"
 
-import RegisterPage from "@/modules/auth/presentation/pages/default/register-page"
+import DefaultRegisterPage from "@/modules/login/presentation/pages/default/DefaultRegisterPage"
 
 export const metadata: Metadata = {
   title: "Create an account",
   description: "Create an account to get started.",
 }
 
-export default RegisterPage
+export default DefaultRegisterPage
