+++ package.json
@@ -61,6 +61,7 @@
     "next": "^14.2.3",
     "next-auth": "^4.24.7",
     "next-themes": "^0.2.1",
+    "nextjs": "^0.0.3",
     "postcss": "^8.4.40",
     "react": "18.3.1",
     "react-day-picker": "^8.10.1",
