--- src/core/models/app.ts
@@ -1,11 +1,0 @@
-import { IApp } from "@/core/models/index"
-
-const app: IApp = {
-  logo: {
-    path: "/images/lis.png",
-    altText: "Next.js Logo",
-    width: 196,
-    height: 32,
-  },
-}
-export default app
