--- src/core/models/index.d.ts
@@ -1,9 +1,0 @@
-export interface IApp {
-  logo: ILogo
-}
-export interface ILogo {
-  path: string
-  altText: string
-  width: number
-  height: number
-}
