+++ src/modules/login/domain/types/auth-config.d.ts
