+++ src/modules/login/presentation/hooks/index.d.ts
@@ -1,0 +1,7 @@
+import { LoginFormValues } from "@/modules/login/config/form-config"
+
+export interface IDefaultLoginHooks {
+  isLoading: boolean
+  loginUser: (values: LoginFormValues) => void
+  getLoginProviderLink: () => void
+}
